+++ net-seller-sheet.php
@@ -25,7 +25,7 @@ define('NSS_PLUGIN_DIR', plugin_dir_path(__FILE__));
 define('NSS_PLUGIN_URL', plugin_dir_url(__FILE__));
 define('NSS_PLUGIN_VERSION', '1.0.0');
 define('NSS_PREFIX', 'nss');
-define('NSS_DB_VERSION', 1);
+define('NSS_DB_VERSION', 2);
 
 // Load Composer autoloader
 require_once NSS_PLUGIN_DIR . 'vendor/autoload.php';